Fix merge errors.
